requirements.txt
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
flake8==6.1.0
frozenlist==1.4.0
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
isort==5.12.0
jmespath==1.0.1
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
minio==7.2.0
multidict==6.0.4
orjson==3.9.10
packaging==23.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
redis==5.0.1
rsa==4.9
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
wrapt==1.16.0
yarl==1.9.2
